refactor: extract business logic from payload classes
