+++ embrace-android-sdk/src/main/java/io/embrace/android/embracesdk/payload/extensions/AnrSampleListExt.kt
@@ -1,0 +1,8 @@
+package io.embrace.android.embracesdk.payload.extensions
+
+import io.embrace.android.embracesdk.payload.AnrSampleList
+
+/**
+ * Retrieves the size of the list.
+ */
+internal fun AnrSampleList.size(): Int = samples.size
